fix circular dependency - help type imports
