+++ packages/renderer/package.json
@@ -45,7 +45,6 @@
     "@polkadot/util": "^12.6.2",
     "@polkadot/util-crypto": "^12.6.2",
     "@substrate/connect": "^1.0.2",
-    "@theme-toggles/react": "^4.1.0",
     "@vitejs/plugin-react-swc": "^3.6.0",
     "@w3ux/extension-assets": "^0.2.6",
     "@w3ux/react-polkicon": "1.2.0-alpha.8",
