+++ packages/renderer/src/config/help.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { HelpItems } from '@ren/renderer/contexts/common/Help/types';
+import type { HelpItems } from '@polkadot-live/types/help';
 
 export const HelpConfig: HelpItems = [
   {
