+++ packages/ui/src/components/Help/index.tsx
@@ -11,7 +11,7 @@ import { ButtonPrimaryInvert } from '../../kits/Buttons/ButtonPrimaryInvert';
 import { CanvasContainer } from '../../kits/Overlay/structure/CanvasContainer';
 import { CanvasScroll } from '../../kits/Overlay/structure/CanvasScroll';
 import { ModalContent } from '../../kits/Overlay/structure/ModalContent';
-import type { HelpProps } from './typess';
+import type { HelpProps } from './types';
 
 export const Help = ({
   status,
