+++ packages/ui/src/components/Help/types.ts
@@ -1,10 +1,7 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type {
-  HelpItem,
-  HelpStatus,
-} from '@ren/renderer/contexts/common/Help/types';
+import type { HelpItem, HelpStatus } from '@polkadot-live/types/help';
 
 export interface HelpProps {
   status: HelpStatus;
